Initial commit containing original ansible-vault code with a few tweaks
This is all the original ansible-vault code with a few tweaks & changes
here and there. I added comments saying when the original code was
changed and left the original code commented out next to the new code.

I also added the "Main" class that lets you exercise the code using a
real ansible-vault executable. The VaultEncryptionHelperTest does NOT
do this, and instead uses mocks. Would be good to find some way to
exercise the code for real.

There are some other things that will need to be addressed, e.g.
using --vault-id when it probably was supposed to be using
--vault-password-file as well as some things that could maybe use
more descriptive/accurate names, e.g. ansibleVaultIdPath in
EncryptionConfiguration really means the path to the password file.

But this commit gets the original code under version control so
I can make changes and see what happens.